Chat Server v0.0.1 Chat.Models.Conversation

Ecto Model for Conversation table.

Functions

add_user(conversation_id, user_id)

Function for adding user to conversation.

Parameters:
conversation_id - ID of conversation. Requires integer
user_id - ID of user. Requires integer

Returns Tuple:
On success: {:ok, user}, where user is added user
On fail: {:error, reason}, where reason is atom representing reason for failing.

change_info(con_id, arg2)

Function for changing conversation details.

Parameters:
con_id - ID of conversation. Requires integer
%{"id" => c_id, "title" => con_name, "avatar" => con_avatar}, where id is conversation ID, title is conversation title and avatar is conversation avatar.

Returns Tuple:
On success: {:ok, conversation}, where conversation is conversation that changed.
On fail: {:error, reason}, where reason is atom representing reason for failing.

get_message_count(con_id)

Function for getting amount of messages in conversation.

Parameters:
con_id - ID of conversation. Requires integer

Returns message count as integer.

get_messages(conversation_id, user_id)

Function for getting messages from conversation for specified user.

We request user_id, becouse we do not want to show user messages from users that he has blocked.

Parameters:
conversation_id - ID of conversation. Requires integer
user_id - ID of user. Requires integer

Returns list of messages.

get_messages(conversation_id, before_id, count, user_id)

Function for getting messages from conversation for specified user.

We request user_id, becouse we do not want to show user messages from users that he has blocked.

Parameters:
conversation_id - ID of conversation. Requires integer
before_id - ID of oldest known message. Requires integer
count - Amount of requested messages. Requires integer
user_id - ID of user. Requires integer

Returns list of messages.

get_users(conversation_id)

Get users of conversations.

Parameters:
conversation_id - ID of conversation. Requires integer

Returns list of users.

in_conversation?(conversation_id, user_id)

Check if user is in conversation?

Parameters:
conversation_id - ID of conversation. Requires integer or binary
user_id - ID of user. Requires integer

Returns boolean.

new(name, users)

Function for creating new conversations.

Parameters:
name - Name of created conversation. Requires binaary value.
users - Users that will take part in conversation. Requires list.

Returns Tuple:
On success: {:ok, conversation}, where conversation is created conversation.
On fail: {:error, reason}, where reason is atom representing reason for failing.

read(conversation_id, user_id, message_id)

Function for setting recently readed message for user on conversation.

Do not allow changing back.

Parameters:
conversation_id - ID of conversation. Requires integer
user_id - ID of user. Requires integer
message_id - ID of message. Requires integer

Returns Tuple:
On success: {:ok, nil}
On fail: {:error, reason}, where reason is atom representing reason for failing.

remove_user(conversation_id, user_id)

Function for removing user from conversation.

Parameters:
conversation_id - ID of conversation. Requires integer
user_id - ID of user. Requires integer

Returns Tuple:
On success: {:ok, conversation}, where conversation is conversation that changed.
On fail: {:error, reason}, where reason is atom representing reason for failing.